refactor(generator): extract Vogen, Interceptor, Agent finders
